refactor(layout): add file ext to relative imports
