+++ packages/layout/src/react/index.js
@@ -1,6 +1,6 @@
-import APageHeadingLayout from './page-heading-layout'
-import AAsideLayout from './aside-layout'
-import AEqualColumnLayout from './equal-column-layout'
+import APageHeadingLayout from './page-heading-layout.js'
+import AAsideLayout from './aside-layout.js'
+import AEqualColumnLayout from './equal-column-layout.js'
 
 export const PageHeadingLayout = APageHeadingLayout
 export const AsideLayout = AAsideLayout
