+++ packages/layout/src/vars/index.js
@@ -1,5 +1,5 @@
-import AAsideLayout from './aside-layout'
-import AEqualColumnLayout from './equal-column-layout'
+import AAsideLayout from './aside-layout.js'
+import AEqualColumnLayout from './equal-column-layout.js'
 
 export const AsideLayout = AAsideLayout
 export const EqualColumnLayout = AEqualColumnLayout
